base/full pypi
