+++ requirements_full.txt
@@ -2,3 +2,4 @@ demucs
 onnx
 onnxruntime
 scenedetect
+-e git+https://github.com/m-bain/whisperx.git
